Added ability to set an orientation.
In the config call you can now pass a third parameter that tells the firmware
the orientation of the ShiftX3.  A value of 0 (the default) indicates the 7-segment
display is on the bottom, a value of 1 indicates it's at the top.  When the device is
inverted and the config is updated accordingly, all linear graph functions, alert led and button locations
retain their user relative ids and behavior.

The only function that doesn't respect the orientation of the device is the set_led api calls.